Temporarily hardcode relative topic links in topics
Relative links break for topics within topics, which
are also `.included` in guides. The LUA `include-files`
filter does not work with the quarto relative paths.
As a result, the working directory changes from `/topics/`
to `/guides/` for the topics included in guides,
resulting in the references for topics referring to the
wrong same folder (that is, `./` now refers to `guides/`
instead of `topics/`). The relative links to guides do not break
because `../guides/` has the same effect regardless of
the working directory. Such a headache!