Update data-documentation.qmd
I couldn't understand why codebook is an exception because it is also partially unstructured metadata - and this is what I see on the image added. So I removed the brackets for a smoother read.
If there is an important thing to say maybe better add it as a full sentence?