Update data-backup.qmd
I rewrote part of the text to make it more general about backup and not too specific about Onedrive. 
I am not sure we really need to go into practical set up of backup here - this is something each tool should cover on its own, in my opinion.
Please feel free to disagree and rewrite this!