Create research-data-and-software-management-policy.qmd
This topic takes the section 'VU General Policies and Regulations' out of the guide [How can you set up research data management from the start?](https://rdm.vu.nl/guides/plan-and-design.html) and updates the information that was presented there.

Once this topic has been included in the Handbook, the section in the above-mentioned guide can be replaced with an include command referring to this new topic.

This topic still contains a reference to the 2020 version of the central policy, but will be replaced as soon as we have the final version of the new policy. I expect this to happen in a few weeks from now.