Create policies-supporting-vision-open-science
This PR creates a new guide that explains which policies, principles and regulations are the foundation of the VU RDSM policy. The target group we have in mind for this page are policy officers and other staff who are involved in policy development, policy implementation, contribute to/maintain policy documents that relate to the RDSM policy, etc.