Update and rename data-publication.qmd to data-publishing.qmd
This is an update of the topic data publishing (previously called 'Data Publication'). It aligns information with the updated RDSM policy. It contains the definition for data publishing and information about the purpose, requirements and how it works in practice.

I deleted the information that was originally there. This information is still available in the guide 'publish and share' and I will adjust this guide later, so that it aligns with the current topic.