Update persistent-identifier.qmd
I added more information about PIDs and how they are used in the research domain.

I noticed that the text is not consistent in capitalising the phrase (Persistent Identifier vs persistent identifier). I don't have a preference for any of those, but it would be nice to be consistent. I'd love to hear if someone has a preference for one of the two.