Update publish-and-share.qmd
This is the first step in updating this guide, to align it with the update RDSM policy, among other things (see also Issue #356).

I changed the title and description, removed the section about the difference between archiving and publishing (which is very different from what we state in the RDSM policy) and added a short new section that addresses making data available as supplementary information, so that I could restructure the section on data publication. I also removed the information about open access/open science, as this is now covered in the topic Open Science.