Create rdsm-terminology.qmd
This is my suggestion for a terminology page on the open handbook. I copied it directly from the RDSM policy and for now I've excluded all of the footnotes. I made this because I think it's helpful to have a version of the terminology that allows for changes and updates more readily than in a policy doc, plus with this method you can link directly to the terms within other documents. You've done this already within the policy PDF, but I don't know how effective that will be when trying to link externally (plus this option allows for iFrames :)