Update pure.qmd
I did some small edits to the text and I removed the include statements for the topics data registration and software registration. Instead, I referred to those instructions with a URL. With the include statements, the topic becomes very long and we run the risk of recursivity (include statements for content that has include statements itself as well).